refactor(gfn): split input protocol modules
